Fixed "Fast refresh only works when a file only exports components". Deleted unused code
